Add new test on consolidation profiles
